refactor: finalize renaming #1040
