+++ src/app/router/RequireAuth.tsx
@@ -1,8 +1,8 @@
 import { Navigate, useLocation } from "react-router-dom";
-import { useAuth } from "~/app/context/AuthContext";
+import { useAccount } from "~/app/context/AccountContext";
 
 function RequireAuth({ children }: { children: JSX.Element }) {
-  const auth = useAuth();
+  const auth = useAccount();
   const location = useLocation();
 
   if (auth.loading) {
